Replacing std::vector with raw arrays in VFEngine::Vector (#10)
* Change #1 - Replacing std::vector with std::arrays for Vector class

* Removing context_memory and datastore member variables from Operator classes, giving direct access to Vector's underlying data

* Removed unecessary comments/ old code

+++ src/memory/include/state.hh
@@ -13,7 +13,7 @@ namespace VFEngine {
 
         void print_debug_info() const;
 
-        static constexpr size_t MAX_VECTOR_SIZE = 1024;
+        static constexpr int32_t MAX_VECTOR_SIZE = 1024;
 
         int32_t _pos;
         int32_t _size;
